requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
pyotp==2.9.0
python-dotenv==1.0.1
Werkzeug==3.1.3
